Move authentication flow caches to tenant space
