+++ components/org.wso2.carbon.identity.oauth/src/main/java/org/wso2/carbon/identity/oauth/cache/OAuthCache.java
@@ -18,14 +18,14 @@
 
 package org.wso2.carbon.identity.oauth.cache;
 
-import org.wso2.carbon.identity.application.common.cache.BaseCache;
+import org.wso2.carbon.identity.application.authentication.framework.cache.AuthenticationBaseCache;
 import org.wso2.carbon.identity.oauth.listener.OAuthCacheRemoveListener;
 import org.wso2.carbon.utils.CarbonUtils;
 
 /**
  * OAuth cache.
  */
-public class OAuthCache extends BaseCache<OAuthCacheKey, CacheEntry> {
+public class OAuthCache extends AuthenticationBaseCache<OAuthCacheKey, CacheEntry> {
 
     private static final String OAUTH_CACHE_NAME = "OAuthCache";
 
@@ -33,7 +33,7 @@ public class OAuthCache extends BaseCache<OAuthCacheKey, CacheEntry> {
 
     private OAuthCache() {
         super(OAUTH_CACHE_NAME);
-        super.addListener(new OAuthCacheRemoveListener());
+//        super.addListener(new OAuthCacheRemoveListener());
     }
 
     public static OAuthCache getInstance() {
