+++ components/org.wso2.carbon.identity.oidc.session/src/main/java/org/wso2/carbon/identity/oidc/session/cache/OIDCBackChannelAuthCodeCacheKey.java
@@ -17,7 +17,7 @@
  */
 package org.wso2.carbon.identity.oidc.session.cache;
 
-import org.wso2.carbon.identity.application.common.cache.CacheKey;
+import org.wso2.carbon.identity.oauth.cache.CacheKey;
 
 /**
  * This class holds the cache key which is AuthorizationCode.
